refactor(changes): extract formatting helpers
